Fix set_custom() and its disable() behavior
There's really nothing to do in the former, setting the root type
to System and setting thissystem to 0 can be done in the discover()
callback. Then we don't need a disable() callback anymore.

Without this, calling set_custom() twice would reset the root type
to Machine when disabling the first set_custom() right after enabling
the second.

This commit was SVN r4716.