Don't explicitly forbid plugins in embedded mod...
Don't explicitly forbid plugins in embedded mode

libltdl embedding is managed in the top configure.ac, so embedders
will have to do their work. he good thing is that it will let them
reuse their existing embedding libltdl if any.

There are likely some stuff to fix in src/Makefile.am to ease that
but we'll see later.

This commit was SVN r4766.