Use InterlockedCompareExchange() to emulate mut...
Use InterlockedCompareExchange() to emulate mutexes on windows

MinGW doesn't offer the pthread ABI, and the external lib pthreads-win32
isn't convenient for building/distributing anyway.

Since we just need one mutex for init()/destroy() and nobody will
ever manipulate multiple topologies in different threads anyway,
we just reimplement a basic mutex on windows using compare-and-swap
and yield (InterlockedCompareExchange() and and SwitchToThread()).

If neither pthread mutex nor InterlockedCompareExchange is found,
configure aborts and asks the user to report to hwloc-devel.

This commit was SVN r4777.