Add a topology-fake.c built as a plugin in make...
Add a topology-fake.c built as a plugin in make check

The plugin is registered as an additional core plugin
(so that it always gets enabled by default).

Its instantiate callback() does nothing but printing something
if HWLOC_DEBUG_FAKE_COMPONENT is set in the environment.
We could grep for this output in a make check test.

It needs "-rpath /nowhere" in LDFLAGS because shared libraries
are only built entirely when placed in lib_LTLIBRARIES
(this fake one is in check_LTLIBRARIES only).

This commit was SVN r4812.