Add flags to backend and component structures
Better now than letter

This may be useful to specify backend requirements.

For instance a backend that only uses the notify_new_object()
callback may not need the topology to be fully connected
when it runs, while a backend with a discover() callback
may need levels and/or children to be fully connected.
We could try to reconnect things only when really needed.

This commit was SVN r4820.