Remove core_ prefix, remove references to class...
Remove core_ prefix, remove references to classes, add hwloc_ prefix and improve the doc

core_*.so becomes hwloc_*.so
xml_*.so becomes hwloc_xml_*.so
Internal structures are not changed.

Remove references to classes in the doc. Remove references to XML specific
backends, except to say that we're not discussing them.

Document (current) component selection system.

This commit was SVN r4853.