Propagate the backend structure into many inter...
Propagate the backend structure into many internal backend callbacks

This is required so that we do not assume anymore than these
backends are first in the list.

There are still some corner cases (osf binding, libpci notification,
and custom check) that need to be fixed.

This commit was SVN r4864.