backends: prepare for enabling multiple backend...
backends: prepare for enabling multiple backends and excluding some

Global backends should not discover() when the topology is
already initialized.

OS backends should return without doing anything (for now) if so.
Except if they want to complete what other backends did earlier
(x86 will do that for FreeBSD, and for the CPUModel info attribute)

This commit was SVN r4953.