components: rename 'core' components into 'disc...
components: rename 'core' components into 'disc' (for discovery)

We may have some other classes of components registered to the core
in the future, the current ones only take care of discovery.

The plugin filenames and component symbol names are still hwloc_foo,
we may want to rename that to hwloc_disc_foo at some point?

This commit was SVN r4978.