components: remove the topology from the instan...
components: remove the topology from the instantiate callback

It was only used so that backend_alloc() puts it in backend->topology,
we now do that in backend_enable() later.

This clarifies the fact that instantiate() should not modify the topology
neither look at it before deciding what to do.
It just fills a backend depending on optional input parameters.

This commit was SVN r4997.